requirements.txt
langchain 
langchain-community 
langchain-google-genai
google-generativeai
gradio
huggingface_hub